requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
SQLAlchemy==2.0.20
PyPDF2==3.0.1
pdfplumber==0.10.2
python-docx==0.8.11
docx2txt==0.8
Pillow==10.0.0
pdf2image==1.16.3
pytesseract==0.3.10
gunicorn==21.2.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
openai==0.28.0
cryptography==41.0.3
pyOpenSSL==23.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
flask-sqlalchemy==2.5.1
numpy==2.1.3
